+++ include/sdf/Geometry.hh
@@ -62,8 +62,21 @@ namespace sdf
     /// \brief Default constructor
     public: Geometry();
 
+    /// \brief Copy constructor
+    /// \param[in] _geometry Geometry to copy.
+    public: Geometry(const Geometry &_geometry);
+
+    /// \brief Move constructor
+    /// \param[in] _geometry Geometry to move.
+    public: Geometry(Geometry &&_geometry);
+
     /// \brief Destructor
     public: virtual ~Geometry();
+
+    /// \brief Assignment operator.
+    /// \param[in] _geometry The geometry to set values from.
+    /// \return *this
+    public: Geometry &operator=(const Geometry &_geometry);
 
     /// \brief Load the geometry based on a element pointer. This is *not* the
     /// usual entry point. Typical usage of the SDF DOM is through the Root